api_yamdb/requirements.txt
# Framework
django==2.2.16
# REST API
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
# pytest & utilits
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
# docker deploy
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.10
python-dotenv==0.21.1
whitenoise==6.3.0 